requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pyserial==3.5
pyautogui==0.9.54
requests==2.32.3
